requirements.txt
Flask==3.0.2
Pillow==10.2.0
gunicorn==21.2.0 